SearchRepositoriesQuery.DataをTranslatorに準拠
そのために子の型の準拠を定義しているファイルはは同じファイルに移行
